Created delete account function

+++ backend/controllers/userController.js
@@ -188,12 +188,17 @@ const deleteUser = asyncHandler(async (req, res) => {
 	const user = await User.findById(req.params.id);
 
 	if (user) {
+		// Admin user can't delete
+		if (user.isAdmin) {
+			res.status(400);
+			throw new Error('Can not Delete Admin User');
+		}
 		await User.deleteOne({ _id: user._id });
 
 		res.status(200).json({ message: 'User Deleted' });
 	} else {
 		res.status(404);
-		throw new Error('This User Not Found');
+		throw new Error('Item Not Found');
 	}
 });
